feat: 일정 목록/내 일정 훅 리팩토링 및 fetch 함수 분리
