+++ src/hooks/itinerary/useMyItineraries.js
@@ -2,8 +2,7 @@ import { useQuery } from "@tanstack/react-query";
 import { auth } from "@/services/firebase";
 import { useAuthState } from "react-firebase-hooks/auth";
 import { QUERY_KEYS } from "@/constants/queryKeys";
-import { collection, getDocs, query, where, orderBy } from "firebase/firestore";
-import { db } from "@/services/firebase";
+import { fetchMyItineraries } from "@/services/itinerary/fetchMyItineraries";
 
 export const useMyItineraries = (options = {}) => {
   const [user, loading] = useAuthState(auth);
@@ -11,23 +10,7 @@ export const useMyItineraries = (options = {}) => {
   return useQuery({
     queryKey: QUERY_KEYS.MY_ITINERARIES(user?.uid),
     enabled: !loading && !!user?.uid && options.enabled !== false,
-    queryFn: async () => {
-      if (!user?.uid) throw new Error("사용자 정보가 없습니다.");
-
-      const q = query(
-        collection(db, "itineraries"),
-        where("userId", "==", user.uid),
-        orderBy("createdAt", "desc")
-      );
-
-      const snapshot = await getDocs(q);
-      return snapshot.docs.map((doc) => ({
-        id: doc.id,
-        ...doc.data(),
-        createdAt: doc.data().createdAt?.toDate() || new Date(),
-        updatedAt: doc.data().updatedAt?.toDate() || new Date(),
-      }));
-    },
+    queryFn: () => fetchMyItineraries(user.uid),
     ...options,
   });
 };
